+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 0.2.0 - TBA
 -----------
 
+- Support for ``MIMEText`` sending. `#25`_
 - Batch emailing implementation. `#12`_
 - Ability to remove headers from email message. `#24`_
 - Improved attachments interface. `#18`_
@@ -22,6 +23,7 @@ Changelog
 - Initial release.
 
 
+.. _#25: https://github.com/FriendlyCoders/postmarker/issues/25
 .. _#24: https://github.com/FriendlyCoders/postmarker/issues/24
 .. _#18: https://github.com/FriendlyCoders/postmarker/issues/18
 .. _#12: https://github.com/FriendlyCoders/postmarker/issues/12